FIX: delete JobValidTypeRempository.java & move validtype  api  to adminDasbord
